Add buttons on Prompt Itemization to view generation data as JSON
Move prompt itemization to own file

Add GENERATION_SUCCESS event emitted by Generate onSuccess

Refactor sendOpenAIRequest

TODO: handle generation responses for multiple swipes

TODO: replace some more of the function coupling with event listeners if possible
