+++ .editorconfig
@@ -5,7 +5,7 @@ end_of_line = lf
 insert_final_newline = true
 trim_trailing_whitespace = true
 
-[*.{js, conf, json, css, less, html}]
+[*.{js,conf,json,css,less,html}]
 charset = utf-8
 indent_style = space
 indent_size = 4
