+++ public/css/mobile-styles.css
@@ -217,6 +217,8 @@
     }
 
     #showRawPrompt,
+    #showGenerationSettings,
+    #showGenerationResponse,
     #copyPromptToClipboard,
     #groupCurrentMemberPopoutButton,
     #summaryExtensionPopoutButton {
