+++ public/scripts/templates/itemizationChat.html
@@ -1,10 +1,31 @@
-<h3 class="flex-container justifyCenter alignitemscenter">
-    Prompt Itemization
-    <div id="showRawPrompt" class="fa-solid fa-square-poll-horizontal menu_button" title="Show Raw Prompt" data-i18n="[title]Show Raw Prompt"></div>
-    <div id="copyPromptToClipboard" class="fa-solid fa-copy menu_button" title="Copy Prompt" data-i18n="[title]Copy Prompt"></div>
-    <div id="diffPrevPrompt" class="fa-solid fa-code-compare menu_button" title="Show Prompt Differences" data-i18n="[title]Show Prompt Differences"></div>
-</h3>
-API/Model Used: {{mainApiFriendlyName}} {{#if apiUsed}}({{apiUsed}}){{/if}} {{#if modelUsed}}&ndash; {{modelUsed}}{{/if}}<br>
+<div class="flex-container justifyCenter alignitemscenter">
+    <h3 class="justifyCenter">Prompt Itemization</h3>
+    <button id="showRawPrompt" class="menu_button" title="Show Raw Prompt" data-i18n="[title]Show Raw Prompt">
+        <i class="fa-solid fa-square-poll-horizontal"></i>
+        <span class="sr-only" data-i18n="Show Raw Prompt">Show Raw Prompt</span>
+    </button>
+    <button id="showGenerationSettings" class="menu_button" title="Show Raw Request" data-i18n="[title]Show Raw Request">
+        <i class="fa-solid fa-code"></i>
+        <i class="fa-regular fa-message"></i>
+        <span class="sr-only" data-i18n="Show Raw Request">Show Raw Request</span>
+    </button>
+    <button id="showGenerationResponse" class="menu_button" title="Show Raw Response" data-i18n="[title]Show Raw Response">
+        <i class="fa-solid fa-code"></i>
+        <i class="fa-regular fa-reply"></i>
+        <span class="sr-only" data-i18n="Show Raw Response">Show Raw Response</span>
+    </button>
+    <button id="copyPromptToClipboard" class="menu_button" title="Copy Prompt" data-i18n="[title]Copy Prompt">
+        <i class="fa-solid fa-copy"></i>
+        <span class="sr-only" data-i18n="Copy Prompt">Copy Prompt</span>
+    </button>
+    <button id="diffPrevPrompt" class="menu_button" title="Show Prompt Differences"
+            data-i18n="[title]Show Prompt Differences">
+        <i class="fa-solid fa-code-compare"></i>
+        <span class="sr-only" data-i18n="Show Prompt Differences">Show Prompt Differences</span>
+    </button>
+</div>
+API/Model Used: {{mainApiFriendlyName}} {{#if apiUsed}}({{apiUsed}}){{/if}} {{#if modelUsed}}&ndash; {{modelUsed}}{{/if}}
+<br>
 Tokenizer: {{selectedTokenizer}}<br>
 <span class="tokenItemizingSubclass">
     Only the white numbers really matter. All numbers are estimates.
@@ -32,27 +53,27 @@ Tokenizer: {{selectedTokenizer}}<br>
                     <div class=""> {{oaiSystemTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Chat Start: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Chat Start:</div>
                     <div class="tokenItemizingSubclass"> {{oaiStartTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Main: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Main:</div>
                     <div class="tokenItemizingSubclass">{{oaiMainTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Post-History: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Post-History:</div>
                     <div class="tokenItemizingSubclass">{{oaiJailbreakTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Auxiliary: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Auxiliary:</div>
                     <div class="tokenItemizingSubclass">{{oaiNsfwTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Nudge: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Nudge:</div>
                     <div class="tokenItemizingSubclass">{{oaiNudgeTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Impersonate: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Impersonate:</div>
                     <div class="tokenItemizingSubclass">{{oaiImpersonateTokens}}</div>
                 </div>
             </div>
@@ -62,7 +83,7 @@ Tokenizer: {{selectedTokenizer}}<br>
                     <div class=""> {{oaiPromptTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Description: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Description:</div>
                     <div class="tokenItemizingSubclass">{{charDescriptionTokens}}</div>
                 </div>
                 <div class="flex-container ">
@@ -70,7 +91,7 @@ Tokenizer: {{selectedTokenizer}}<br>
                     <div class="tokenItemizingSubclass"> {{charPersonalityTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Scenario: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Scenario:</div>
                     <div class="tokenItemizingSubclass">{{scenarioTextTokens}}</div>
                 </div>
                 <div class="flex-container ">
@@ -96,7 +117,7 @@ Tokenizer: {{selectedTokenizer}}<br>
                     <div class="">{{allAnchorsTokens}}</div>
                 </div>
                 <div class="flex-container ">
-                    <div class=" flex1 tokenItemizingSubclass">-- Summarize: </div>
+                    <div class=" flex1 tokenItemizingSubclass">-- Summarize:</div>
                     <div class="tokenItemizingSubclass">{{summarizeStringTokens}}</div>
                 </div>
                 <div class="flex-container ">
